pyproject.toml
[tool.poetry]
name = "discord-ai"
version = "0.1.0"
description = ""
authors = ["sadeshmukh <thesahildeshmukh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nextcord = "^2.6.0"
python-dotenv = "^1.0.1"
ollama = "^0.3.2"
google-generativeai = "^0.7.2"
anthropic = "^0.34.1"
openai = "^1.43.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
